my_project/requirements.txt
Django==1.10.1
django-debug-toolbar==1.6
cx_Oracle==5.2.1
gunicorn==19.6.0
whitenoise==3.2.1
